store player direction
this is direction they are facing. we can't base this on speed because
we need to know what direction a player was facing last after they've
stopped. since we can deterine a players previous x or y location by
their speed, we no longer need playerPrevX or playerPrevY so I've used
that memory space for direction.

right now there is still a piece of memory that can used for another
player property where playerPrevY used to be.