queue spears up so that you can only have 3 at a time on screen
utilize those extra spear data structures so that whenever you spawn a
spear, it shifts the spears down a queue and despawns the latest spear
that was thrown.

also despawn a spear when it falls off the bottom of the screen. this
messes with the queue by adding inactive spears, but should be such a
rare occurence that it might not ever be noticed.