start the game with a menu screen
the art is still a placeholder and doesn't contain instructions - press
'1' to start the game.

I modified the `currentZone` convention to start at the menu screen, so
every zone has been bumped up one - the caves are now currentZone 1
instead of 0 and so forth. this adds in a couple of extra instructions
to restart the game at currentZone 1 and needing to subtract 1 from
currentZone when choosing monster types (as it's a jump offset that
starts from 0). setting the menu zone to currentZone 3 would avoid these
issues but puts it at a weird place in the code. might be worth it
though.